requirements.txt
absl-py==0.15.0
astunparse==1.6.3
blessings==1.7
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
distlib==0.3.4
filelock==3.6.0
flatbuffers==1.12
fonttools==4.33.3
gast==0.4.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpustat==0.6.0
grpcio==1.34.1
h5py==3.1.0
idna==3.3
imageio==2.19.0
imgaug==0.4.0
importlib-metadata==4.11.3
joblib==1.1.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
Markdown==3.3.6
matplotlib==3.5.2
networkx==2.8
numpy==1.19.3
nvidia-ml-py3==7.352.0
oauthlib==3.2.0
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.3
Pillow==9.1.0
platformdirs==2.5.1
protobuf==3.20.1
psutil==5.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pylon==0.4.4
pyparsing==3.0.8
pypylon==1.8.0
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scikit-learn==0.24.1
scipy==1.8.0
Shapely==1.8.1.post1
six==1.15.0
startrek==0.3.8
tcp==0.3.8
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2022.4.28
typing-extensions==3.7.4.3
urllib3==1.26.9
utility==1.0
virtualenv==20.14.0
Wand==0.6.7
Werkzeug==2.1.2
wrapt==1.12.1
zipp==3.8.0
